requirements.txt
Flask
pyTelegramBotAPI
pyngrok
SQLAlchemy
requests
aiohttp
supabase-py
APScheduler
